requirements.txt
Django~=4.1.0
django-taggit~=3.0.0
markdown~=3.4.1
psycopg2-binary~=2.9.3
python-dotenv
newsapi~=0.2.7
pydantic~=2.9.2